+++ changelog.md
@@ -4,7 +4,8 @@ v0.3.2
 - Improved unit tests for Ohm's Law calculator.
 - Converted CalcVars object from Dictionary to List, closes #23.
 - Removed input variable to validator function, #25.
-- Added correct validators to IPC2221A track current calculator, closes #24.
+- Added correct validators to IPC2221A Track Current calculator, closes #24.
+- Added visual diagram to IPC2221A Track Current calculator, which includes a top PCB layer which disappears if "External" is selected as the track layer.
 
 v0.3.1
 ------